Upgrade rust edition (#449)
* fix: rust edition to 2021

* feat: update to polkadot v0.9.13 to fix ChainSpecExtention macro issue

* fix: remove light-client support

light-client is experiencing major change in the upstream. It's also not supported for para-chains yet

* fix: wiring of the new slash parameter disable_strategy

The parameter is not yet consumed properly in our slash computation. The proper support will come in https://github.com/nodlecode/chain-workspace/issues/95

* fix: resolving para-id from cli. Use default para-id (1000) with "eden-local" and "eden-dev".

The id/para-id is not an explicit run command for a parachain anymore. It's assumed the parachain id is part of the extended chain spec. So for example no need to specify both "eden" and its id. "eden" should have a known id.

* fix(ci): specify the main-net wss port explicitly.